Move ltp instructions to its own file
For the sake of organization and readability, move the ltp instructions
out of the example folder README.md and put it in its own file. Also
updates the description based on utilizing a shared ltp source
directory.

Also introduce a script that allows us to run a subset of ltp tests
without having to install the whole suite.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>