Add a kernel module example
Adds an example kernel module that actually does things. This is meant
to serve as an example for how I can setup testcases to targetedly hit
my code utilizing exported symbols.

The flow here is:
1) Build the kernel w/ external module support.
2) Write a module that sets up the scenario for my new features.
3) Hotplug the module in to test the feature, then remove it. If an
   OOPS happens, or we know we're leaking memory that affects our code,
   we can reboot the QEMU instance and be fine.
4) Repeat step 2 as necessary (only doing step 1 to change my new
   feature).

Also pass the "printk.synchronous=1" argument to the command line to
make printk() more consistent.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>