Add a userspace example for syscalls
This is an mmap() example. Other syscall usages can be figured out
through the man pages. All we need to do is boot into the kernel,
compile this file, then run the binary. Run the binary in the
background and uncomment the sleep() call to have our mmap()
allocation persist in the background.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>