Add an example for manipulating a character device
Provides an example for manipulating a character device. A block device
is similar: "register_blkdev()" along with a "struct
block_device_operations" instead.

Whenever userspace calls an implemented function, the kernel lets the
module handle it based on the implementation. We must create a file to
be treated as the 'bridge' for data between kernelspace and userspace.

I.e. For pagecache testing, we can open a file in userspace, pass that
file descriptor using an ioctl, then operate on that file & its mapping.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>