Introduce QEMU command line argument
For QEMU testing of various architectures, we introduce a command line
arguemnt that lets us override the default qemu-kvm command. There are 2
relevant commands, one that matches the emulated architecture (based on
the QEMU default naming scheme), and one that replaces the entire command.

This is for usecases when we do not have kvm support, or want to do a
full virtual emulation.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>