Add serial parameter to QEMU command
We can boot syzkaller BzImages directly after downloading them. In order
to access the machine with our files, we must recreate the 'nographic'
parameter for QEMU. We do that by passing -serial stdio to QEMU. This
allows us to see the dmesg output as the syzkaller image is booting, so
we know when we can sucessfully ssh into the machine.

There's probably some kernel-config related reason that syskaller
kernels show no output to the default DISPLAY/MONITOR parameters, but I
haven't looked into those details yet.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>