qemu_test: Change kernel option functionality
It doesn't make a lot of sense to have part of the kernel pathway
hardcoded. It also becomes a hassle to try to autofill pathways. For
that reason, change the functionality of the '-k' option to take a
pathway to the kernel directory. It will still search for the arch/
directory to find the kernel corresponding to our architecture.

The '-K' option remains unchanged, and still expects the absolute path
to the entire bzImage itself.

Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>