Example Scripts: fix module makefile
The module makefile was forcing the Makefile to 'guess' and autogenerate
kernel pathways for module output. This made it impossible to move a
directory after generating external modules for it once.

Furthermore, it hardcoded the module output location, which prevented
the script from copying the compiled module to our shared directory.

Fix this by directly exporting the desired environment variables prior
to the request.

Fixes: 5d9feaf5 ("External Module Support")
Signed-off-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>